fix: binding result 사용 answer 중복생성 수정 - #54

+++ src/main/java/com/peer/missionpeerflow/controller/QuestionController.java
@@ -8,10 +8,13 @@ import com.peer.missionpeerflow.exception.message.SuccessMessage;
 import com.peer.missionpeerflow.service.QuestionService;
 
 import lombok.RequiredArgsConstructor;
+import org.hibernate.QueryParameterException;
+import org.hibernate.hql.internal.QueryExecutionRequestException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -24,26 +27,34 @@ public class QuestionController {
     public final QuestionService questionService;
 
     @PostMapping("/v1/question")
-    public ResponseEntity<Object> create(@Valid @RequestBody QuestionCreateDTO questionCreateDTO) {
+    public ResponseEntity<Object> create(@Valid @RequestBody QuestionCreateDTO questionCreateDTO, BindingResult bindingResult) {
+        if (bindingResult.hasErrors())
+            throw new QueryParameterException("Query parameter is invalid");
         questionService.creat(questionCreateDTO);
         return ResponseEntity.status(HttpStatus.OK).body(SuccessMessage.of("Question created successfully"));
     }
 
     @GetMapping("/v1/question/{questionid}")
-    public ResponseEntity<Object> detail(Model model, @PathVariable(name = "questionid") Long id) {
+    public ResponseEntity<Object> detail(Model model, @PathVariable(name = "questionid") Long id, BindingResult bindingResult) {
+        if (bindingResult.hasErrors())
+            throw new QueryParameterException("Query parameter is invalid");
         QuestionDetailDTO questionDetailDTO = this.questionService.getQuestionDetail(id);
         model.addAttribute("questionDetail", questionDetailDTO);
         return ResponseEntity.status(HttpStatus.OK).body(questionDetailDTO);
     }
 
     @PutMapping("/v1/question/{questionid}")
-    public ResponseEntity<Object> modify(@Valid @RequestBody QuestionModifyDTO questionModifyDTO, @PathVariable(name = "questionid") Long id) {
+    public ResponseEntity<Object> modify(@Valid @RequestBody QuestionModifyDTO questionModifyDTO, @PathVariable(name = "questionid") Long id, BindingResult bindingResult) {
+        if (bindingResult.hasErrors())
+            throw new QueryParameterException("Query parameter is invalid");
         questionService.modify(questionModifyDTO, id);
         return ResponseEntity.status(HttpStatus.OK).body(SuccessMessage.of("question modified successfully"));
     }
 
     @PostMapping("/v1/question/{questionid}")
-    public ResponseEntity<Object> delete(@Valid @RequestBody QuestionDeleteDTO questionDeleteDTO, @PathVariable(name = "questionid") Long id) {
+    public ResponseEntity<Object> delete(@Valid @RequestBody QuestionDeleteDTO questionDeleteDTO, @PathVariable(name = "questionid") Long id, BindingResult bindingResult) {
+        if (bindingResult.hasErrors())
+            throw new QueryParameterException("Query parameter is invalid");
         questionService.delete(questionDeleteDTO, id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).body(SuccessMessage.of("question deleted successfully"));
     }
